requirements.txt
telethon
python-decouple
pytz